Extract React Router Prompt customizations
